requirements.txt
pandas
numpy
requests
python-dotenv
matplotlib
seaborn
plotly
scikit-learn
xgboost
shap
apscheduler
